+++ src/structures/gdt.rs
@@ -1,7 +1,10 @@
 //! Types for the Global Descriptor Table and segment selectors.
 
-use crate::structures::tss::TaskStateSegment;
 use crate::PrivilegeLevel;
+use crate::{
+    structures::{tss::TaskStateSegment, DescriptorTablePointer},
+    VirtAddr,
+};
 use bit_field::BitField;
 use bitflags::bitflags;
 use core::fmt;
@@ -105,33 +108,69 @@ impl GlobalDescriptorTable {
         }
     }
 
-    /// Adds the given segment descriptor to the GDT, returning the segment selector.
+    /// Forms a GDT from a slice of `u64`.
     ///
-    /// Panics if the GDT has no free entries left.
+    /// # Safety
+    ///
+    /// * The user must make sure that the entries are well formed
+    /// * The provided slice **must not be larger than 8 items** (only up to the first 8 will be observed.)
     #[inline]
-    pub fn add_entry(&mut self, entry: Descriptor) -> SegmentSelector {
-        let index = match entry {
-            Descriptor::UserSegment(value) => self.push(value),
-            Descriptor::SystemSegment(value_low, value_high) => {
-                let index = self.push(value_low);
-                self.push(value_high);
-                index
-            }
-        };
+    #[cfg(feature = "const_fn")]
+    pub const unsafe fn from_raw_slice(slice: &[u64]) -> GlobalDescriptorTable {
+        assert!(
+            slice.len() <= 8,
+            "initializing a GDT from a slice requires it to be **at most** 8 elements."
+        );
+        let next_free = slice.len();
 
-        let rpl = match entry {
-            Descriptor::UserSegment(value) => {
-                if DescriptorFlags::from_bits_truncate(value).contains(DescriptorFlags::DPL_RING_3)
-                {
-                    PrivilegeLevel::Ring3
-                } else {
-                    PrivilegeLevel::Ring0
+        let mut table = [0; 8];
+        let mut idx = 0;
+
+        while idx != next_free {
+            table[idx] = slice[idx];
+            idx += 1;
+        }
+
+        GlobalDescriptorTable { table, next_free }
+    }
+
+    /// Get a reference to the internal table.
+    ///
+    /// The resulting slice may contain system descriptors, which span two `u64`s.
+    #[inline]
+    pub fn as_raw_slice(&self) -> &[u64] {
+        &self.table[..self.next_free]
+    }
+
+    const_fn! {
+        /// Adds the given segment descriptor to the GDT, returning the segment selector.
+        ///
+        /// Panics if the GDT has no free entries left.
+        #[inline]
+        pub fn add_entry(&mut self, entry: Descriptor) -> SegmentSelector {
+            let index = match entry {
+                Descriptor::UserSegment(value) => self.push(value),
+                Descriptor::SystemSegment(value_low, value_high) => {
+                    let index = self.push(value_low);
+                    self.push(value_high);
+                    index
                 }
-            }
-            Descriptor::SystemSegment(_, _) => PrivilegeLevel::Ring0,
-        };
+            };
 
-        SegmentSelector::new(index as u16, rpl)
+            let rpl = match entry {
+                Descriptor::UserSegment(value) => {
+                    if DescriptorFlags::from_bits_truncate(value).contains(DescriptorFlags::DPL_RING_3)
+                    {
+                        PrivilegeLevel::Ring3
+                    } else {
+                        PrivilegeLevel::Ring0
+                    }
+                }
+                Descriptor::SystemSegment(_, _) => PrivilegeLevel::Ring0,
+            };
+
+            SegmentSelector::new(index as u16, rpl)
+        }
     }
 
     /// Loads the GDT in the CPU using the `lgdt` instruction. This does **not** alter any of the
@@ -142,26 +181,51 @@ impl GlobalDescriptorTable {
     #[cfg(feature = "instructions")]
     #[inline]
     pub fn load(&'static self) {
-        use crate::instructions::tables::{lgdt, DescriptorTablePointer};
-        use core::mem::size_of;
-
-        let ptr = DescriptorTablePointer {
-            base: self.table.as_ptr() as u64,
-            limit: (self.table.len() * size_of::<u64>() - 1) as u16,
-        };
-
-        unsafe { lgdt(&ptr) };
+        // SAFETY: static lifetime ensures no modification after loading.
+        unsafe { self.load_unsafe() };
     }
 
+    /// Loads the GDT in the CPU using the `lgdt` instruction. This does **not** alter any of the
+    /// segment registers; you **must** (re)load them yourself using [the appropriate
+    /// functions](crate::instructions::segmentation):
+    /// [load_ss](crate::instructions::segmentation::load_ss),
+    /// [set_cs](crate::instructions::segmentation::set_cs).
+    ///
+    /// # Safety
+    ///
+    /// Unlike `load` this function will not impose a static lifetime constraint
+    /// this means its up to the user to ensure that there will be no modifications
+    /// after loading and that the GDT will live for as long as it's loaded.
+    ///
+    #[cfg(feature = "instructions")]
     #[inline]
-    fn push(&mut self, value: u64) -> usize {
-        if self.next_free < self.table.len() {
-            let index = self.next_free;
-            self.table[index] = value;
-            self.next_free += 1;
-            index
-        } else {
-            panic!("GDT full");
+    pub unsafe fn load_unsafe(&self) {
+        use crate::instructions::tables::lgdt;
+        lgdt(&self.pointer());
+    }
+
+    const_fn! {
+        #[inline]
+        fn push(&mut self, value: u64) -> usize {
+            if self.next_free < self.table.len() {
+                let index = self.next_free;
+                self.table[index] = value;
+                self.next_free += 1;
+                index
+            } else {
+                panic!("GDT full");
+            }
+        }
+    }
+
+    /// Creates the descriptor pointer for this table. This pointer can only be
+    /// safely used if the table is never modified or destroyed while in use.
+    #[cfg(feature = "instructions")]
+    fn pointer(&self) -> DescriptorTablePointer {
+        use core::mem::size_of;
+        DescriptorTablePointer {
+            base: VirtAddr::new(self.table.as_ptr() as u64),
+            limit: (self.next_free * size_of::<u64>() - 1) as u16,
         }
     }
 }